requirements.txt
requests
pytz
aiosqlite
pynacl
ffmpeg
discord.py[voice]
